Stat Mech and Graphing Implementation
This commit comes with a full implementation of the Stat Mech model that
Chesson was working on. Additionally, this includes the framework to
begin 3-D modeling of state space! This has a lot of potential to graph
a variety of different situations we want to see.

TODO:
Want to add command line interface