Graphing Overhall/Stability Calculator
In this commit, I went ahead and added the methods to calculate the
stability of the different models, as well as combining Somil's code
with the master copy to keep everything in one piece. In order to
calculate the stability of the graph, I went off the assumption that
Volume is held constant for a point, and that we can observe the
change in energy with respect to temperature. This is Cv, and if it is
greater than 0, we are at a stable point.

To overhaul the graphing functionality, I created a graphing class which
takes in the labels for the specific scenario. In addition, I went ahead
and made it graph each point at a time so that I can individually change
the labels depending on the stability around that point. I know that the
documentation for it is pretty sloppy, but it does what it needs to do
well so far.

TODO:
Add functionality to easily adjust/display the state space of the
Entropy plotting, as well as updated methods to calculate the entropy
from them.