Pressure Updates
Our last version of the scripts did not accurately account for the
changing state variables, and instead assumed that they would all be
constant asides from the user input ones. This commit ensures that this
is no longer the case for the pressure.

Additionally, this added support for metastable points and improved the
README to account for the pressure changes.